Add files via upload
CreditCard.java - This program checks whether a credit card number is valid or invalid using the Luhn algorithm, which is commonly used by credit card companies as a simple integrity test. It verifies that the credit card number complies with the standard format by performing a series of steps, including doubling digits, summing them, and checking the last digit for correctness. 
AlphabetWarGame- This program, AlphabetWarGame, simulates a battle between the "left" and "right" sides based on specific letters in the input string. It assigns points to each side: letters 'w', 'p', 'b', and 's' give points to the left side, while 'm', 'q', 'd', and 'z' give points to the right side. After calculating the total score for each side, the program declares a winner or calls for a rematch if the scores are tied.
task2_program_1.java- This program, task2_program_1, finds the top k most frequent numbers in a predefined integer array and displays them in descending order of frequency.

This program, ShareTrade, calculates the maximum possible profit from at most two stock transactions. Each transaction consists of buying and then selling shares, and the goal is to maximize the profit.